backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
sqlalchemy==2.0.23
pyodbc==5.0.1
pymongo==4.6.0
azure-storage-blob==12.19.0
azure-identity==1.15.0
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
pydantic==2.5.0
pydantic-settings==2.1.0
alembic==1.13.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
jinja2==3.1.2
reportlab==4.0.7
pillow==10.1.0
motor==3.3.2
email-validator==2.1.0
bcrypt==4.1.2
